Big changes Merge spfa-generator into it Use commander to parse command User chalk to show logger Put the code into directories

+++ src/commands/clean.js
@@ -22,6 +22,8 @@
 
 const fs = require("fs");
 const { join } = require("path");
+const check = require("../utils/check");
+const logger = require("../utils/logger");
 
 // Remove the "public" directory
 function rmdir(dir, callback) {
@@ -53,13 +55,29 @@ function rmdir(dir, callback) {
 }
 
 function clean() {
-    rmdir(join(process.cwd(), "public"), function () {
-        console.log("Cleaned!");
-        fs.mkdir(join(process.cwd(), "public"), function (err) {
-            fs.mkdir(join(process.cwd(), "public/post"), function (err) {});
-            fs.mkdir(join(process.cwd(), "public/lib"), function (err) {});
+    check(process.cwd())
+        .then(function (path) {
+            if (!path) {
+                logger.error("Please initialize first!");
+                return;
+            }
+            process.chdir(path);
+            rmdir(join(process.cwd(), "public"), function () {
+                logger.info("Cleaned!");
+                fs.mkdir(join(process.cwd(), "public"), function (err) {
+                    fs.mkdir(join(process.cwd(), "public/post"), function (
+                        err
+                    ) {});
+                    fs.mkdir(join(process.cwd(), "public/lib"), function (
+                        err
+                    ) {});
+                });
+            });
+        })
+        .catch(function (err) {
+            logger.error(err);
+            process.exit(0);
         });
-    });
 }
 
 // Export the function
